商品削除機能是正1

+++ app/controllers/items_controller.rb
@@ -38,8 +38,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    if item.destroy
+    if current_user.id != item.user_id
       redirect_to root_path
+    else
+      if item.destroy
+        redirect_to root_path
+      end
     end
   end
